+++ WS2012R2/lisa/tools/aws/tests/run_sysbench.sh
@@ -31,17 +31,22 @@
 #       3. Runs sysbench
 #       4. Prepares results
 #
-#       No optional parameters needed
-#
 ########################################################################
 LOG_FILE=/tmp/summary.log
 function LogMsg() {
     echo $(date "+%a %b %d %T %Y") : ${1} >> ${LOG_FILE}
 }
 
+if [ $# -lt 1 ]; then
+    echo -e "\nUsage:\n$0 device"
+    exit 1
+fi
+
+DISK="$1"
 MODES=(seqwr seqrewr seqrd rndrd rndwr rndrw)
 THREADS=(1 2 4 8 16 32 64 128 256 512 1024)
 IOS=(4 8 16 32)
+EXTRA="--file-total-size=184G --max-requests=0 --max-time=300 --file-extra-flags=dsync --file-fsync-freq=0"
 
 if [ -e /tmp/summary.log ]; then
     rm -rf /tmp/summary.log
@@ -50,15 +55,9 @@ fi
 sudo apt-get update >> ${LOG_FILE}
 sudo apt-get -y install libaio1 sysstat zip sysbench >> ${LOG_FILE}
 
-cd /tmp
-mkdir -p /tmp/sysbench_fileio
-sudo sysbench --test=fileio cleanup >> ${LOG_FILE}
-sudo sysbench --test=fileio prepare >> ${LOG_FILE}
-
 function fileio ()
 {
     LogMsg " Testing sysbench fileio mode=$1 ios=$2"K" threads=$3."
-    EXTRA="--file-total-size=134G --file-extra-flags=dsync --file-fsync-freq=0 --max-requests=0 --max-time=300"
     iostat -x -d 1 900 2>&1 > /tmp/sysbench_fileio/$1_$2"K"_$3_iostat.diskio.log &
     vmstat 1 900       2>&1 > /tmp/sysbench_fileio/$1_$2"K"_$3_vmstat.memory.cpu.log &
 
@@ -72,6 +71,25 @@ function fileio ()
     sudo pkill -f vmstat
 }
 
+mkdir -p /tmp/sysbench_fileio
+
+if [[ ${DISK} == *"xvd"* ]]
+then
+    sudo mkfs.ext4 ${DISK}
+    sudo mkdir /stor
+    sudo mount ${DISK} /stor
+    sudo chmod 777 /stor
+    cd /stor
+elif [[ ${DISK} == *"md"* ]]
+then
+    sudo chmod 777 /raid
+    cd /raid
+else
+    LogMsg "Failed to identify disk type for ${DISK}."
+    exit 70
+fi
+
+sudo sysbench --test=fileio --file-total-size=184G prepare >> ${LOG_FILE}
 for mode in "${MODES[@]}"
 do
     for io in "${IOS[@]}"
@@ -80,18 +98,14 @@ do
         do
             fileio ${mode} ${io} ${thread}
         done
+        sleep 10
     done
+    sleep 10
 done
-
-sudo sysbench --test=fileio cleanup
+sudo sysbench --test=fileio --file-total-size=184G cleanup >> ${LOG_FILE}
 
 LogMsg "Kernel Version : `uname -r` "
 
 cd /tmp
 zip -r sysbench.zip . -i sysbench_fileio/* >> ${LOG_FILE}
 zip -r sysbench.zip . -i summary.log >> ${LOG_FILE}
-
-function finish {
-  LogMsg "Execution ended with exit code $?"
-}
-trap finish EXIT
